requirements.txt
jupyter
ipykernel
keyring
pandas
requests
google-genai
dotenv